Rest Api without DB was added plus some file location changes

+++ labs3sem/OS_Labs/OS_LAB5/Client/Client.cpp
@@ -10,6 +10,7 @@ const int Exit = 5;
 const int Success = 6;
 const int Fail = 7;
 
+
 void PrintEmp(const Employee& e)
 {
     std::cout << "ID: " << e.num << ", name: " << e.name << ", hours: " << e.hours << std::endl;
@@ -17,16 +18,12 @@ void PrintEmp(const Employee& e)
 
 int main() 
 {
-
-    HANDLE hPipe = INVALID_HANDLE_VALUE;
+    HANDLE hPipe;
     const wchar_t* pipeName = L"\\\\.\\pipe\\MyPipe";
     char c;
     while (true) 
     {
-        hPipe = CreateFileW(
-            pipeName,
-            GENERIC_READ | GENERIC_WRITE,
-            0, NULL, OPEN_EXISTING, 0, NULL);
+        hPipe = CreateFileW(pipeName, GENERIC_READ | GENERIC_WRITE, 0, NULL, OPEN_EXISTING, 0, NULL);
 
         if (hPipe == INVALID_HANDLE_VALUE)
         {
@@ -46,10 +43,9 @@ int main()
 
     while (true)
     {
-        std::cout << "1. Read Record\n2. Modify Record\n3. Exit\nYour choice: ";
+        std::cout << "1. Read\n2. Modify\n3. Exit\nYour choice: ";
         int choice;
         std::cin >> choice;
-
         DWORD bytesWritten, bytesRead;
 
         if (choice == 1)
@@ -87,7 +83,6 @@ int main()
             int id;
             std::cout << "Enter employee ID: ";
             std::cin >> id;
-
             int cmd = WriteStart;
             WriteFile(hPipe, &cmd, sizeof(int), &bytesWritten, NULL);
             WriteFile(hPipe, &id, sizeof(int), &bytesWritten, NULL);
@@ -96,8 +91,11 @@ int main()
             if (ReadFile(hPipe, &resp, sizeof(int), &bytesRead, NULL) && resp == Success)
             {
                 Employee e;
+                bool isCorrect = false;
                 ReadFile(hPipe, &e, sizeof(Employee), &bytesRead, NULL);
 
+                while (!isCorrect)
+                {
                 std::cout << "EDITING..."<<std::endl;
                 std::cout << "Current data: ";
                 PrintEmp(e);
@@ -105,16 +103,29 @@ int main()
                 std::cin >> e.name;
                 std::cout << "Enter new hours: ";
                 std::cin >> e.hours;
+                
+                std::string choice;
+                std::cout << "Are you sure the data is correct?(yes/no) ";
+                std::cin >> choice;
+                 if (choice == "yes")
+                  {
+                        std::string finishCmd;
+                        std::cout << "Type any letter to send data: ";
+                        std::cin >> finishCmd;
 
-               
-                char finishCmd;
-                std::cout << "Type any letter to send data: ";
-                std::cin >> finishCmd;
 
-                WriteFile(hPipe, &e, sizeof(Employee), &bytesWritten, NULL);
-                cmd = WriteEnd;
-                WriteFile(hPipe, &cmd, sizeof(int), &bytesWritten, NULL);
-                std::cout << "Everything was sent to server. You may continue" << std::endl;
+                        WriteFile(hPipe, &e, sizeof(Employee), &bytesWritten, NULL);
+                        cmd = WriteEnd;
+                        WriteFile(hPipe, &cmd, sizeof(int), &bytesWritten, NULL);
+                        std::cout << "Everything was sent to server. You may continue" << std::endl;
+                        isCorrect = true;
+                  }
+                else if (choice == "no")
+                {
+                     std::cout << "Let's try again" << std::endl;
+                }
+
+                }
             }
             else 
             {
